backend/requirements.txt
uvicorn[standard]
fastapi
pymongo
redis
python-dotenv
psycopg[binary]
psycopg_pool
httpx